Add some __init__.py's to make the vendor directory and the markdown directory actual modules so that we can import them in python.
Second, I've made some modifications from the markdown.py, the reason for which you can read here:

http://blog.reddit.com/2009/09/we-had-some-bugs-and-it-hurt-us.html

The patch is not exactly the same as the one the guys at Reddit created since I am not using Pylon. Instead a SECRET key now has to be added to the config, which is different from anyone elses key just to make sure that nothing nasty happens.

http://code.reddit.com/changeset/1f1f0606f5b6bf14a0db55a28cfd03e1e42e3550

The original markdown.py is taken from the web.py sourcecode available at:

http://github.com/webpy/webpy/blob/master/tools/markdown.py



git-svn-id: http://svn.0x58.com/repos/BertJWRegeer/Projects/appblog/trunk@124 3d59c550-c999-dd11-821d-0030482bb012