Getting ready for the big commit at r135.
There are several changes, first the order of the properties has changed, while this means that nothing has really changed it affects the order in which the fields show up for the form. This is a bug in the way djangoforms handlers its different properties. This has been updated in a later version of Django which is not included with AppEngine.

Also, create a form for the entry which can then be used to add new entries.

Slug has become a MagicProperty which is calculated when title is set, this allows us to satisfy the requirement for it to required, and for it to be calculated after the fact. Currently it is calculating the length of the title, debug only.



git-svn-id: http://svn.0x58.com/repos/BertJWRegeer/Projects/appblog/trunk@134 3d59c550-c999-dd11-821d-0030482bb012